Move all tools.py functions to submodule temporal and update notebook
